FIRST COMMIT - Adding existing source code, ConfigFile and plot tools.
Added all developed code for the 3D Navier-Stokes FSI solver.
In this state, the solver gives good results for uniform flow tests and channel flow tests.
Initial- and Boundary conditions for these cases are hard-coded in functions.
Now, the next step is to implement the simplified 3D IBM (as in Matlab code) for this N-S solver.