Symmetry BC seems to be fixed.
No more velocity component normal to the symmetry planes.
Changed back to injection (?) at inlet and outlet. Still using ghost
nodes at periodic and symmetry. The final change that fixed it was to
mirror the primitives instead of the conserved vars.

Also made some changes to how grid spacings are computed, and how to
find spatial position in the mesh.